Wolt_Assignment_Pritish/requirements.txt
Flask==2.0.2
requests==2.26.0
pytest==6.2.5
python-dateutil==2.8.2